requirements.txt
Django==4.1.1
djangorestframework==3.13.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-jazzmin==2.6.0
django-cleanup==7.0.0
django-solo==2.1.0

python-dotenv==1.0.0
Pillow==10.4.0
gunicorn==20.1.0
whitenoise==6.3.0

drf-yasg==1.21.5
coreapi==2.3.3

# Base de datos
psycopg2-binary==2.9.9
dj-database-url==1.2.0

channels==4.0.0
channels-redis==4.1.0
django-notifications-hq==1.8.3
asgiref==3.7.2

django-rest-passwordreset==1.3.0

requests==2.31.0

